fix: use r3f renderloop of smoothscrollbar only from default import target
