+++ src/components/R3FSmoothScrollbar.tsx
@@ -1,7 +1,10 @@
-import React from 'react'
+import React, { forwardRef } from 'react'
 import { SmoothScrollbar } from '../scrollbar/SmoothScrollbar'
+import { ISmoothScrollbar } from '../scrollbar/SmoothScrollbar.d'
 import { addEffect, invalidate } from '@react-three/fiber'
 
-export default function R3FSmoothScrollbar(props: any) {
-  return <SmoothScrollbar invalidate={invalidate} addEffect={addEffect} {...props} />
+function R3FSmoothScrollbar(props: ISmoothScrollbar, ref: any) {
+  return <SmoothScrollbar ref={ref} invalidate={invalidate} addEffect={addEffect} {...props} />
 }
+
+export default forwardRef<any, ISmoothScrollbar>(R3FSmoothScrollbar)
